requirements.txt
google-api-python-client==2.111.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
anthropic==0.18.1
sqlalchemy==2.0.25
streamlit==1.31.0
python-dotenv==1.0.0
pandas==2.2.0
